auto_bid_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
asgiref==3.7.2
beautifulsoup4==4.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
compressed_rtf==1.0.6
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
IMAPClient==2.1.0
lxml==5.1.0
oauthlib==3.2.2
olefile==0.47
openai==1.14.1
pdfminer==20191125
pdfminer.six==20191110
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-pptx==0.6.23
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
textract==1.6.5
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
xlrd==1.2.0
XlsxWriter==3.2.0
